Collocate organizational signatories
There are 3 of the form `[individual] on behalf of [organization]` which I left alone as I'm not sure whether they are properly individual or organizational signatories. But aside from that, I did a pass through to find the organizations scattered throughout the list and put them with the bulk of the organizations, to aid in later list alphabetization/organization.